refactor(Name, NameExceptionCode): 패키지 위치 name -> global로 이동
