+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -14,7 +14,7 @@ import roomescape.global.exception.model.RoomEscapeException;
 import roomescape.member.domain.Member;
 import roomescape.reservation.exception.ReservationExceptionCode;
 import roomescape.theme.domain.Theme;
-import roomescape.time.domain.ReservationTime;
+import roomescape.reservationtime.domain.ReservationTime;
 
 @Entity
 public class Reservation {
