+++ src/main/java/roomescape/theme/domain/Theme.java
@@ -6,7 +6,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import roomescape.name.domain.Name;
+import roomescape.global.domain.Name;
 
 @Entity
 public class Theme {
@@ -26,7 +26,7 @@ public class Theme {
     public Theme() {
     }
 
-    private Theme(long id, Name name, String description, String thumbnail) {
+    public Theme(long id, Name name, String description, String thumbnail) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -45,10 +45,6 @@ public class Theme {
         return new Theme(name, description, thumbnail);
     }
 
-    public static Theme saveThemeFrom(long id) {
-        return new Theme(id, null, null, null);
-    }
-
     public long getId() {
         return id;
     }
@@ -64,9 +60,4 @@ public class Theme {
     public String getThumbnail() {
         return thumbnail;
     }
-
-    public void setIdOnSave(long id) {
-        this.id = id;
-    }
-
 }
